% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgalh-package.R
\docType{package}
\name{cgalh-package}
\alias{cgalh}
\alias{cgalh-package}
\title{Provides the 'C++' header-only library of CGAL}
\description{
'CGAL' is a C++ library that aims to provide easy access to efficient and reliable algorithms in computational geometry. Since its version 4, 'CGAL' can be used as standalone header-only library and is available under a double GPL-3|LGPL license. <https://www.cgal.org/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/dickoa/cgalh}
  \item Report bugs at \url{https://gitlab.com/dickoa/cgalh/issues}
}

}
