% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgalh-package.R
\docType{package}
\name{cgalh-package}
\alias{cgalh}
\alias{cgalh-package}
\title{Provides CGAL 'C++' header-only library.}
\description{
CGAL is a C++ library that aims to provide easy access to efficient and reliable algorithms in computational geometry.
}
\examples{
\dontrun{

# Include a dependency to cgalh inside a cpp file
// [[Rcpp::depends(cgalh)]]
include <Rcpp.h>
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/dickoa/cgalh}
  \item Report bugs at \url{https://gitlab.com/dickoa/cgalh/issues}
}

}
\author{
\strong{Maintainer}: Ahmadou Dicko \email{mail@ahmadoudicko.com}

}
